+++ src/cpp/MultiLayerPerceptron.h
@@ -21,7 +21,7 @@ class MultiLayerPerceptron {
  private:
   static const unsigned int MAX_TRIAL = 1000; // 学習上限回数
   constexpr static const double MAX_GAP = 0.1; // 損失関数の出力がこの値以下になれば学習をスキップする
-  int num_thread = android_getCpuCount(); // Androidデバイスのプロセッサのコア数
+  unsigned long num_thread = (unsigned long)android_getCpuCount(); // Androidデバイスのプロセッサのコア数
 
   string mlp_params = "";
